Add vec_xl_len and vec_xst_len
They are powerpc64-only.
